Make ownership of proxies explicit, see #9
this adds the infrastructure to avoid memory leaks. However, it does not
seem to work very well. While some things get released, such as
`prepared_setup`, others don't, namely the things that are bound to the
lifetime of `js_exports`. Do we never release the `js_exports` or is the
vue3-sfc-compiler to blame here?